# internal/web/static/openapi.json
{
    "swagger": "2.0",
    "info": {
        "description": "The API the Tunnel Manager screens are built on. Every path\nbelow sits under /api and needs a session, except POST\n/api/login and GET /api/setup.\n\nLogging in sets two cookies. Over HTTPS they are named\n__Host-tm_session and __Host-tm_csrf, and over plain HTTP they\nare tm_session and tm_csrf. A browser sends them back on its\nown, so \"Try it out\" on this page works as soon as you are\nlogged in to the screens in the same browser.\n\nEvery POST, PUT and DELETE also needs the token of the session\nin an X-CSRF-Token header. POST /api/login answers with it in\ndata.csrf_token; paste it into Authorize above and this page\nsends it. A GET needs no token.\n\nEvery answer has the same shape: {\"success\":true,\"data\":...}\nor {\"success\":false,\"error\":\"...\"}. An answer that says no\ncarries error_code, which names the refusal and is what a\nscript should match on, and error_args, the values its English\nsentence was written with.",
        "title": "Tunnel Manager API",
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        },
        "version": "3.8.3"
    },
    "basePath": "/api",
    "paths": {
...
        "/certificate": {
            "get": {
                "description": "Fingerprint, subject, issuer, the names it covers, the validity and the days left. The private key is never in the answer.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "certificate"
                ],
                "summary": "The certificate being served",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.certificateView"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "api_https_enabled is off, so no certificate is in use",
                        "schema": {
                            "$ref": "#/definitions/api.errorBody"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "CSRFToken": []
                    }
                ],
                "description": "Takes cert_pem and key_pem, stores them and serves them from the next connection on. cert_pem may be a chain, with the server certificate first and the intermediates behind it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "certificate"
                ],
                "summary": "Register a certificate of your own",
                "parameters": [
                    {
                        "description": "The certificate and its private key, as PEM",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.certificateInstall"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.certificateReplaced"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "The PEM could not be read, or the key does not match the certificate",
                        "schema": {
                            "$ref": "#/definitions/api.errorBody"
                        }
                    },
                    "409": {
                        "description": "api_https_enabled is off, so no certificate is in use",
                        "schema": {
                            "$ref": "#/definitions/api.errorBody"
                        }
                    }
                }
            }
        },
        "/certificate/renew": {
            "post": {
                "security": [
                    {
                        "CSRFToken": []
                    }
                ],
                "description": "A replacement takes effect on the next connection and not on the ones already open, so the answer to this arrives over the old certificate.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "certificate"
                ],
                "summary": "Make another self-signed certificate and serve it from the next connection on",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.certificateReplaced"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "api_https_enabled is off, so no certificate is in use",
                        "schema": {
                            "$ref": "#/definitions/api.errorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.certificateInstall": {
            "type": "object",
            "properties": {
                "cert_pem": {
                    "type": "string"
                },
                "key_pem": {
                    "type": "string"
                }
            }
        },
        "api.certificateReplaced": {
            "type": "object",
            "properties": {
                "certificate": {
                    "$ref": "#/definitions/api.certificateView"
                },
                "note": {
                    "type": "string"
                },
                "note_code": {
                    "description": "NoteCode names the note, so that a screen says it in the language it is\ndrawn in instead of showing the English above.",
                    "type": "string"
                },
                "previous_fingerprint_sha256": {
                    "type": "string"
                },
                "warning": {
                    "description": "Warning is what is wrong with a certificate that was stored anyway, or \"\"\nwhen there is nothing to say. Install fills it for a certificate whose\nvalidity has not started yet.",
                    "type": "string"
                },
                "warning_args": {
                    "$ref": "#/definitions/api.textArgs"
                },
                "warning_code": {
                    "description": "WarningCode names that warning and WarningArgs holds the value written\ninto it. Both are left out where there is no warning, and the code is\nleft out for a warning this file has no name for, which is the case the\nscreen shows the English sentence for.",
                    "type": "string"
                }
            }
        },
        "api.certificateView": {
            "type": "object",
            "properties": {
                "cert_pem": {
                    "description": "CertPEM is here so that an operator can take the self-signed certificate\ninto the trust store of a machine without going to the database for it.\nIt is what the server hands to every client during the handshake anyway.",
                    "type": "string"
                },
                "days_remaining": {
                    "type": "integer"
                },
                "fingerprint_sha256": {
                    "type": "string"
                },
                "hosts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "issuer": {
                    "type": "string"
                },
                "not_after": {
                    "type": "string"
                },
                "not_before": {
                    "type": "string"
                },
                "self_signed": {
                    "type": "boolean"
                },
                "subject": {
                    "type": "string"
                }
            }
        },
        "api.errorArgs": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "api.errorBody": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "error_args": {
                    "$ref": "#/definitions/api.errorArgs"
                },
                "error_code": {
                    "$ref": "#/definitions/api.errorCode"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.textArgs": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CSRFToken": {
            "description": "The token POST /api/login answers with, in\ndata.csrf_token. Required on every POST, PUT\nand DELETE. The session cookie goes with it and\nis sent by the browser on its own.",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    }
}